+++ src/common.h
@@ -4,7 +4,7 @@
 #ifndef COMMON_H
 #define COMMON_H
 
-#include <stdint.h>
+#include <cstdint>
 #include <QString>
 
 class QUrl;